requirements.txt
flask==3.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2
pyodbc==5.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
